\name{numbers2words, numbers2words_cap1, numbers2words_cap2}
\alias{numbers2words}
\title{Numbers to words}
\usage{
numbers2words(x)
}
\description{
modification of function taken from John Fox http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
Converts numbers to words. cap1 capitilises 11-90. cap2 capitilises 1-10
}
\examples{
numbers2words(34)
}
