\name{numbers2words, numbers2words_cap1, numbers2words_cap2}
\alias{numbers2words}
\title{Numbers to words}
\usage{
numbers2words(x)
}
\description{
modification of function included in Global Environment
Converts numbers to words. cap1 capitilises 11-90. cap2 capitilises 1-10
}
\examples{
numbers2words(34)
}
